# docs/openapi.yaml
openapi: 3.0.3
info:
  title: apple-bridge
  description: macOS EventKit HTTP API — access Apple Reminders and Calendars
  version: 1.0.0
servers:
  - url: http://127.0.0.1:23487
    description: Local server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key created via `apple-bridge keys create`
  schemas:
    AlarmInput:
      type: object
      properties:
        relativeOffset:
          type: number
          description: Seconds relative to event/due date (negative = before)
          example: -3600
        absoluteDate:
          type: string
          format: date-time
          example: "2026-04-15T09:00:00Z"
    AlarmResponse:
      type: object
      properties:
        relativeOffset:
          type: number
          nullable: true
        absoluteDate:
          type: string
          format: date-time
          nullable: true
    RecurrenceRuleInput:
      type: object
      required: [frequency, interval]
      properties:
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        interval:
          type: integer
          example: 1
        daysOfTheWeek:
          type: array
          items:
            type: string
            enum: [MO, TU, WE, TH, FR, SA, SU]
        daysOfTheMonth:
          type: array
          items:
            type: integer
        monthsOfTheYear:
          type: array
          items:
            type: integer
        end:
          type: object
          properties:
            date:
              type: string
              format: date-time
            count:
              type: integer
    RecurrenceRuleResponse:
      type: object
      properties:
        frequency:
          type: string
        interval:
          type: integer
        daysOfTheWeek:
          type: array
          items:
            type: string
          nullable: true
        daysOfTheMonth:
          type: array
          items:
            type: integer
          nullable: true
        monthsOfTheYear:
          type: array
          items:
            type: integer
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        endCount:
          type: integer
          nullable: true
    LocationInput:
      type: object
      required: [latitude, longitude]
      properties:
        title:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        radius:
          type: number
          description: Radius in meters (default 100)
        proximity:
          type: string
          enum: [enter, leave]
    LocationResponse:
      type: object
      properties:
        title:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        proximity:
          type: string
    CreateReminderRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: Buy groceries
        notes:
          type: string
        listId:
          type: string
          description: Reminder list identifier. Omit to use default list.
        dueDate:
          type: string
          format: date-time
        priority:
          type: integer
          description: "0 = none, 1 = high, 5 = medium, 9 = low"
        alarms:
          type: array
          items:
            $ref: '#/components/schemas/AlarmInput'
        recurrenceRules:
          type: array
          items:
            $ref: '#/components/schemas/RecurrenceRuleInput'
        location:
          $ref: '#/components/schemas/LocationInput'
        url:
          type: string
    UpdateReminderRequest:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
        listId:
          type: string
        dueDate:
          type: string
          format: date-time
        priority:
          type: integer
        alarms:
          type: array
          items:
            $ref: '#/components/schemas/AlarmInput'
        recurrenceRules:
          type: array
          items:
            $ref: '#/components/schemas/RecurrenceRuleInput'
        location:
          $ref: '#/components/schemas/LocationInput'
        url:
          type: string
    ReminderResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        notes:
          type: string
          nullable: true
        listId:
          type: string
        listTitle:
          type: string
        isCompleted:
          type: boolean
        completionDate:
          type: string
          format: date-time
          nullable: true
        dueDate:
          type: string
          format: date-time
          nullable: true
        priority:
          type: integer
        alarms:
          type: array
          items:
            $ref: '#/components/schemas/AlarmResponse'
        recurrenceRules:
          type: array
          items:
            $ref: '#/components/schemas/RecurrenceRuleResponse'
        location:
          $ref: '#/components/schemas/LocationResponse'
        url:
          type: string
          nullable: true
        creationDate:
          type: string
          format: date-time
          nullable: true
        lastModifiedDate:
          type: string
          format: date-time
          nullable: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
paths:
...
  /reminders:
    get:
      summary: List reminders
      tags: [Reminders]
      parameters:
        - name: listId
          in: query
          schema:
            type: string
        - name: completed
          in: query
          schema:
            type: boolean
        - name: dueBefore
          in: query
          schema:
            type: string
            format: date-time
        - name: dueAfter
          in: query
          schema:
            type: string
            format: date-time
        - name: priority
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of reminders
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReminderResponse'
                  count:
                    type: integer
    post:
      summary: Create a reminder
      tags: [Reminders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReminderRequest'
      responses:
        '201':
          description: Reminder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reminders/{id}:
    get:
      summary: Get a reminder
      tags: [Reminders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reminder found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update a reminder (full)
      tags: [Reminders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          description: Reminder updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: Not found
    patch:
      summary: Update a reminder (partial)
      tags: [Reminders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReminderRequest'
      responses:
        '200':
          description: Reminder updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: Not found
    delete:
      summary: Delete a reminder
      tags: [Reminders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /reminders/{id}/complete:
    post:
      summary: Mark reminder complete
      tags: [Reminders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reminder completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: Not found
  /reminders/{id}/uncomplete:
    post:
      summary: Mark reminder incomplete
      tags: [Reminders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reminder marked incomplete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderResponse'
        '404':
          description: Not found
...